Added some checks to ensure that edges maintain legal templates and template switching behavior.  Also added edge template parameterization as well as all of the remaining types of edge templates.
One known bug remains: When toggling the arrow on an edge, the template gets removed correctly if it is no longer legal, but does not draw correctly until another refresh occurs.  Not sure how to fix this.

I would also like to spend some time tweaking the visuals (maybe change the style of the buttons currently acting as controls) and adding any remaining tool-tips that might be helpful.